Move shared enum model classes to pip-data-models
