+++ build.gradle
@@ -186,7 +186,7 @@ dependencies {
   implementation group: 'net.logstash.logback', name: 'logstash-logback-encoder', version: '7.3'
 
   implementation group: 'uk.gov.service.notify', name: 'notifications-java-client', version: '3.19.1-RELEASE'
-  implementation group: 'com.github.hmcts', name: 'pip-data-models', version: '1.4.2', {
+  implementation group: 'com.github.hmcts', name: 'pip-data-models', version: '1.4.5', {
     exclude group: 'org.springframework.boot', module: 'spring-boot-starter-data-jpa'
   }
