--- src/main/java/uk/gov/hmcts/reform/pip/publication/services/authentication/roles/IsAdmin.java
@@ -1,18 +1,0 @@
-package uk.gov.hmcts.reform.pip.publication.services.authentication.roles;
-
-import org.springframework.security.access.prepost.PreAuthorize;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Method that captures the admin role.
- */
-@Target({ElementType.TYPE, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-@PreAuthorize("hasAuthority('APPROLE_api.request.admin')")
-public @interface IsAdmin {
-
-}
