--- src/main/java/uk/gov/hmcts/reform/pip/publication/services/models/external/ArtefactType.java
@@ -1,8 +1,0 @@
-package uk.gov.hmcts.reform.pip.publication.services.models.external;
-
-/**
- * An enum which represents the Artefact type.
- */
-public enum ArtefactType {
-    LIST, JUDGEMENTS_AND_OUTCOMES, GENERAL_PUBLICATION
-}
