--- src/main/java/uk/gov/hmcts/reform/pip/publication/services/models/external/FileType.java
@@ -1,9 +1,0 @@
-package uk.gov.hmcts.reform.pip.publication.services.models.external;
-
-/**
- * Enum used in the map of files returned from channel management.
- */
-public enum FileType {
-    EXCEL,
-    PDF
-}
