--- src/main/java/uk/gov/hmcts/reform/pip/publication/services/models/external/Language.java
@@ -1,8 +1,0 @@
-package uk.gov.hmcts.reform.pip.publication.services.models.external;
-
-/**
- * A class which defines the language of the publication.
- */
-public enum Language {
-    ENGLISH, WELSH, BI_LINGUAL
-}
