--- src/main/java/uk/gov/hmcts/reform/pip/publication/services/models/external/Location.java
@@ -1,20 +1,0 @@
-package uk.gov.hmcts.reform.pip.publication.services.models.external;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class Location {
-
-    private Integer locationId;
-
-    private String name;
-
-    private String welshName;
-
-    private List<String> jurisdiction;
-
-    private List<String> region;
-
-}
