--- src/main/java/uk/gov/hmcts/reform/pip/publication/services/models/external/Sensitivity.java
@@ -1,8 +1,0 @@
-package uk.gov.hmcts.reform.pip.publication.services.models.external;
-
-/**
- * A class which defines the sensitivity of the publication.
- */
-public enum Sensitivity {
-    PUBLIC, CLASSIFIED, PRIVATE
-}
